remove use of namespaces
