+++ client/.eslintrc.js
@@ -30,7 +30,6 @@ module.exports = {
         ],
         '@typescript-eslint/require-await': 'error',
         '@typescript-eslint/await-thenable': 'error',
-        '@typescript-eslint/no-namespace': 'off',
         '@typescript-eslint/no-explicit-any': 'off',
         '@typescript-eslint/no-var-requires': 'off',
         '@typescript-eslint/no-unused-vars': 'off',
