+++ client/package.json
@@ -11,8 +11,8 @@
         "test:device": "npm run build && mocha dist/RokuDevice.spec.js dist/OnDeviceComponent.spec.js",
         "test:all": "npm run build && mocha dist/**/*.spec.js",
         "buildConfigSchema": "typescript-json-schema src/types/ConfigOptions.ts ConfigOptions --out rta-config.schema.json --noExtraProps --refs false --required --ignoreErrors",
-        "buildRequestTypesSchema": "typescript-json-schema src/types/OnDeviceComponentRequest.ts ODC.RequestType --out requestTypes.schema.json --ignoreErrors",
-        "buildRequestArgsSchema": "typescript-json-schema src/types/OnDeviceComponentRequest.ts ODC.RequestArgs --propOrder --out requestArgs.schema.json --ignoreErrors",
+        "buildRequestTypesSchema": "typescript-json-schema src/types/OnDeviceComponent.ts RequestType --out requestTypes.schema.json --ignoreErrors",
+        "buildRequestArgsSchema": "typescript-json-schema src/types/OnDeviceComponent.ts RequestArgs --propOrder --out requestArgs.schema.json --ignoreErrors",
         "buildRelease": "npx ts-node scripts/buildRelease.ts",
         "buildRelease:beta": "npx ts-node scripts/buildRelease.ts --beta",
         "buildRelease:dev": "npx ts-node scripts/buildRelease.ts --dev",
