+++ client/src/index.ts
@@ -23,4 +23,4 @@ export { Suitest, suitest };
 
 export * from './types/ActiveAppResponse';
 export * from './types/ConfigOptions';
-export * from './types/OnDeviceComponentRequest';
+export * from './types/OnDeviceComponent';
